New empty timetable created
