+++ src/components/Timetable.tsx
@@ -1,181 +1,155 @@
-import React, { useState, useEffect, useRef } from 'react';
-import { Table, TableHead, TableBody, TableRow, TableCell } from '@mui/material';
-import type { Course } from '../types/Course';
-import type { ClassData } from '../types/ClassData';
+import React from 'react';
+import { Course } from '@/types/Course';
+import { ClassData } from '@/types/ClassData';
+import { Table, TableRow, TableCell } from '@mui/material';
+
+const numberToPercent = (num: number): string => {
+  return num.toString() + '%';
+};
+
+interface TimeColumnProps {
+  startTime: number;
+  rows: number;
+}
+
+const TimeColumn: React.FC<TimeColumnProps> = ({ startTime, rows }) => {
+  const hours: number[] = [...Array(rows)].map((_, i) => i + startTime);
+  return (
+    <div>
+      <Table style={{ height: '100%', display: 'flex', width: '50px', flexDirection: 'column' }}>
+        {hours.map(h => {
+          return (
+            <TableRow key={h} style={{ flexGrow: 2, display: 'flex', justifyContent: 'center', alignItems: 'center', flexBasis: '0%' }}>
+              <TableCell style={{ borderBottom: '0px', padding: '0px' }}>
+                <strong style={{ color: 'rgb(105, 112, 122)' }}>{h * 100}</strong>
+              </TableCell>
+            </TableRow>
+          );
+        })}
+        <TableRow style={{ flexGrow: 1 }}>
+          <TableCell style={{ padding: '0px', border: '0px' }}>
+            <strong style={{ color: 'rgb(105, 112, 122)' }}></strong>
+          </TableCell>
+        </TableRow>
+      </Table>
+    </div>
+  );
+};
+
+interface ScheduleProps {
+  rows: number;
+}
+
+const Schedule: React.FC<ScheduleProps> = ({ rows }) => {
+  return (
+    <div style={{ height: '100%' }}>
+      <Table style={{ height: '100%', display: 'flex', flexDirection: 'column' }}>
+        <TableRow style={{ flexGrow: 1, display: 'flex', flexDirection: 'row', flexBasis: '0%' }}>
+          {['MON', 'TUE', 'WED', 'THU', 'FRI'].map(day => {
+            const border = '1px solid rgba(224, 224, 224)';
+            const borderLeft = day === 'MON' ? '1px solid rgba(224, 224, 224)' : '0px';
+            const topLeftRadius = day === 'MON' ? '10px' : '0px';
+            const topRightRadius = day === 'FRI' ? '10px' : '0px';
+            return (
+              <TableCell
+                key={day}
+                style={{
+                  borderTopLeftRadius: topLeftRadius,
+                  borderTopRightRadius: topRightRadius,
+                  flexGrow: 1,
+                  display: 'flex',
+                  flexBasis: '0%',
+                  alignItems: 'center',
+                  justifyContent: 'center',
+                  padding: '0px',
+                  borderRight: border,
+                  borderLeft: borderLeft,
+                  borderBottom: '0px',
+                  borderTop: border,
+                }}
+              >
+                <strong style={{ color: 'rgb(105, 112, 122)' }}>{day}</strong>
+              </TableCell>
+            );
+          })}
+        </TableRow>
+        {[...Array(rows)].map((_, i) => {
+          const border = '1px solid rgba(224, 224, 224)';
+          const borderBottom = i === rows - 1 ? border : '0px';
+          const backgroundColor = i % 2 === 0 ? '#efefef' : 'white';
+          return (
+            <TableRow style={{ flexGrow: 2, display: 'flex', flexDirection: 'row', flexBasis: '0%' }} key={i}>
+              {[...Array(5)].map((_, n) => {
+                const borderLeft = n === 0 ? '1px solid rgba(224, 224, 224)' : '0px';
+                const bottomLeftRadius = i === rows - 1 && n === 0 ? '10px' : '0px';
+                const bottomRightRadius = i === rows - 1 && n === 4 ? '10px' : '0px';
+                return (
+                  <TableCell
+                    key={n}
+                    style={{
+                      borderBottomLeftRadius: bottomLeftRadius,
+                      borderBottomRightRadius: bottomRightRadius,
+                      display: 'flex',
+                      flexBasis: '0%',
+                      backgroundColor: backgroundColor,
+                      borderBottom: borderBottom,
+                      borderRight: border,
+                      borderLeft: borderLeft,
+                      padding: '0px',
+                      flexGrow: 1,
+                    }}
+                  ></TableCell>
+                );
+              })}
+            </TableRow>
+          );
+        })}
+      </Table>
+    </div>
+  );
+};
+
+interface ClassCardsProps {
+  courses: Course[];
+  startTime: number;
+  rows: number;
+}
+
+const ClassCards: React.FC<ClassCardsProps> = ({ courses, startTime, rows }) => {
+  return (
+    <div style={{ position: 'absolute', top: 42, left: 0, width: '100%', height: '100%' }}>
+      <Table>
+        <TableRow>
+          {[...Array(5)].map((_, i) => {
+            return <TableCell key={i} style={{ height: '100%' }}></TableCell>;
+          })}
+        </TableRow>
+      </Table>
+    </div>
+  );
+};
 
 interface TimetableProps {
   courses: Course[];
 }
 
 const Timetable: React.FC<TimetableProps> = ({ courses }) => {
-  const [selectedCells, setSelectedCells] = useState<ClassData[]>([]);
-  const tableRef = useRef<HTMLDivElement>(null);
+  // Default start and end of days: 10AM - 6PM
+  const startTime = 10;
+  const endTime = 18;
+  const rows = endTime - startTime;
 
-  const days = ['MON', 'TUE', 'WED', 'THU', 'FRI'];
-
-  const groupClassDataByLessonType = (timetableData: ClassData[]) => {
-    const groupedData: { [lessonType: string]: [ClassData[], string] } = {};
-
-    timetableData.forEach(data => {
-      const { lessonType } = data;
-      if (!groupedData[lessonType]) {
-        groupedData[lessonType] = [[data], data.classNo];
-      } else {
-        groupedData[lessonType][0].push(data);
-      }
-    });
-
-    return groupedData;
-  };
-
-  const groupedDataByCourse: { [courseName: string]: { [lessonType: string]: [ClassData[], string] } } = {};
-  courses.forEach(course => {
-    groupedDataByCourse[course.courseName] = groupClassDataByLessonType(course.timetableData);
-  });
-
-  const findClassForHour = (course: Course, day: string, hour: number) => {
-    const lessons = groupedDataByCourse[course.courseName];
-    for (const lessonType in lessons) {
-      const [classData, classNo] = lessons[lessonType];
-      const chosenClass = classData.find(data => data.classNo === classNo);
-      const isMatch =
-        chosenClass?.day === day &&
-        parseInt(chosenClass?.startTime as string) / 100 <= hour &&
-        parseInt(chosenClass?.endTime as string) / 100 > hour;
-      if (isMatch) {
-        return chosenClass;
-      }
-    }
-  };
-
-  const handleCellClick = (course: Course, day: string, hour: number) => {
-    const newSelectedCells: ClassData[] = [];
-    const lessons = groupedDataByCourse[course.courseName];
-    for (const lessonType in lessons) {
-      const [classData] = lessons[lessonType];
-      const matchingClass = classData.find(
-        data => data.day === day && parseInt(data.startTime) / 100 <= hour && parseInt(data.endTime) / 100 > hour,
-      );
-      if (matchingClass) {
-        newSelectedCells.push(matchingClass);
-      }
-    }
-
-    // Toggle selection: if the cell is already selected, deselect it, otherwise select it
-    if (selectedCells.some(cell => newSelectedCells.some(newCell => cell === newCell))) {
-      setSelectedCells([]);
-    } else {
-      setSelectedCells(newSelectedCells);
-    }
-  };
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      let isOutside = true;
-      selectedCells.forEach(selectedCell => {
-        const cellId = `${selectedCell.day}-${selectedCell.startTime}-${selectedCell.endTime}`;
-        if ((event.target as HTMLElement).id === cellId) {
-          isOutside = false;
-        }
-      });
-
-      if (isOutside) {
-        setSelectedCells([]);
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [selectedCells]);
-
-  const lighterColor = (color: string) => {
-    // Convert hex color to RGB
-    const hex = color.replace('#', '');
-    let r = parseInt(hex.substring(0, 2), 16);
-    let g = parseInt(hex.substring(2, 4), 16);
-    let b = parseInt(hex.substring(4, 6), 16);
-
-    // Calculate lighter color (reduce darkness)
-    const amount = 0.3; // Adjust the value to control the lightness
-    r = Math.round(Math.min(255, r + 255 * amount));
-    g = Math.round(Math.min(255, g + 255 * amount));
-    b = Math.round(Math.min(255, b + 255 * amount));
-
-    // Convert back to hex
-    const lighterHex = (r * 65536 + g * 256 + b).toString(16);
-    return `#${lighterHex.padStart(6, '0')}`;
+  const getClassesFromCourses = (courses: Course[]): ClassData[] => {
+    return courses.map(course => course.timetableData).reduce((acc, val) => acc.concat(val), []);
   };
 
   return (
-    <div ref={tableRef} style={{ overflowX: 'auto' }}>
-      <Table style={{ tableLayout: 'fixed', minWidth: '600px' }}>
-        <TableHead>
-          <TableRow>
-            <TableCell style={{ width: '60px', fontSize: '12px', borderBottom: '0px' }}></TableCell>
-            {days.map(day => (
-              <TableCell key={day} style={{ width: '100px', fontSize: '12px', textAlign: 'center' }}>
-                {day}
-              </TableCell>
-            ))}
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {[...Array(13)].map((_, index) => {
-            const hour = index + 8;
-            const time = hour < 12 ? `${hour} AM` : `${hour === 12 ? 12 : hour - 12} PM`;
-            const color = hour % 2 ? '#eeeeee' : 'white';
-            console.log(index);
-            const bottomBorder = index === 12 ? '1px solid rgpa(224, 224, 224)' : '0px';
-
-            return (
-              <TableRow key={hour} style={{ height: '50px' }}>
-                <TableCell
-                  style={{ fontSize: '12px', borderBottom: '0px', borderRight: '1px solid rgba(224, 224, 224)', textAlign: 'right' }}
-                >
-                  {time}
-                </TableCell>
-                {days.map(day => {
-                  const matchingCourse = courses.find(course => findClassForHour(course, day, hour));
-                  if (matchingCourse) {
-                    const matchingClass = findClassForHour(matchingCourse, day, hour);
-                    const isSelected = selectedCells.some(cell => cell === matchingClass);
-                    return (
-                      <TableCell
-                        key={`${day}-${hour}`}
-                        id={`${day}-${matchingClass?.startTime}-${matchingClass?.endTime}`}
-                        onClick={() => handleCellClick(matchingCourse, day, hour)}
-                        style={{
-                          backgroundColor: isSelected ? lighterColor(matchingCourse.color) : matchingCourse.color,
-                          width: '100px',
-                          fontSize: '10px',
-                          borderBottom: bottomBorder,
-                          borderRight: '1px solid rgba(224, 224, 224)',
-                          cursor: 'pointer',
-                        }}
-                      >{`${matchingCourse.courseName} ${matchingClass?.lessonType} ${matchingClass?.classNo}`}</TableCell>
-                    );
-                  } else {
-                    return (
-                      <TableCell
-                        key={`${day}-${hour}`}
-                        style={{
-                          width: '100px',
-                          fontSize: '10px',
-                          borderBottom: bottomBorder,
-                          borderRight: '1px solid rgba(224, 224, 224)',
-                          backgroundColor: color,
-                        }}
-                      ></TableCell>
-                    );
-                  }
-                })}
-              </TableRow>
-            );
-          })}
-        </TableBody>
-      </Table>
+    <div style={{ display: 'flex', margin: '10px', height: 800, width: '99%' }}>
+      <TimeColumn startTime={startTime} rows={rows} />
+      <div style={{ flexGrow: 1, position: 'relative' }}>
+        <Schedule rows={rows} />
+        {/* <ClassCards courses={courses} startTime={startTime} rows={rows} /> */}
+      </div>
     </div>
   );
 };
